Rename `ConnectionPool` to `Node`
