+++ rust/connlib/snownet/tests/lib.rs
@@ -1,5 +1,5 @@
 use boringtun::x25519::StaticSecret;
-use snownet::{ClientConnectionPool, Event};
+use snownet::{ClientNode, Event};
 use std::{
     collections::HashSet,
     time::{Duration, Instant},
@@ -10,7 +10,7 @@ fn connection_times_out_after_10_seconds() {
     let start = Instant::now();
 
     let mut alice =
-        ClientConnectionPool::<u64>::new(StaticSecret::random_from_rng(rand::thread_rng()), start);
+        ClientNode::<u64>::new(StaticSecret::random_from_rng(rand::thread_rng()), start);
 
     let _ = alice.new_connection(1, HashSet::new(), HashSet::new());
     alice.handle_timeout(start + Duration::from_secs(10));
