+++ rust/snownet-tests/src/main.rs
@@ -12,10 +12,7 @@ use futures::{channel::mpsc, future::BoxFuture, FutureExt, SinkExt, StreamExt};
 use pnet_packet::{ip::IpNextHeaderProtocols, ipv4::Ipv4Packet};
 use redis::{aio::MultiplexedConnection, AsyncCommands};
 use secrecy::{ExposeSecret as _, Secret};
-use snownet::{
-    Answer, ClientConnectionPool, ConnectionPool, Credentials, IpPacket, Offer,
-    ServerConnectionPool,
-};
+use snownet::{Answer, ClientNode, Credentials, IpPacket, Node, Offer, ServerNode};
 use tokio::{io::ReadBuf, net::UdpSocket};
 use tracing_subscriber::EnvFilter;
 
@@ -79,7 +76,7 @@ async fn main() -> Result<()> {
 
     match role {
         Role::Dialer => {
-            let mut pool = ClientConnectionPool::<u64>::new(private_key, Instant::now());
+            let mut pool = ClientNode::<u64>::new(private_key, Instant::now());
             pool.add_local_interface(socket_addr);
 
             let offer = pool.new_connection(
@@ -162,7 +159,7 @@ async fn main() -> Result<()> {
             }
         }
         Role::Listener => {
-            let mut pool = ServerConnectionPool::<u64>::new(private_key, Instant::now());
+            let mut pool = ServerNode::<u64>::new(private_key, Instant::now());
             pool.add_local_interface(socket_addr);
 
             let offer = redis_connection
@@ -335,7 +332,7 @@ impl FromStr for Role {
 
 struct Eventloop<T> {
     socket: UdpSocket,
-    pool: ConnectionPool<T, u64>,
+    pool: Node<T, u64>,
     timeout: BoxFuture<'static, Instant>,
     candidate_rx: mpsc::Receiver<wire::Candidate>,
     read_buffer: Box<[u8; MAX_UDP_SIZE]>,
@@ -345,7 +342,7 @@ struct Eventloop<T> {
 impl<T> Eventloop<T> {
     fn new(
         socket: UdpSocket,
-        pool: ConnectionPool<T, u64>,
+        pool: Node<T, u64>,
         candidate_rx: mpsc::Receiver<wire::Candidate>,
     ) -> Self {
         Self {
